Skeleton for navigation through (MainMenuScreen --> CharacterSelectionScreen --> StageSelectionScreen --> Stage)
This commit sets up the basic structure for screen navigation.
Components and controllers now include a simple function that can be tied to each screen's button to move to the next screen (though the actual button-function connections are not yet made).

Currently, none of these screens are connected via interfaces, so they do not know each other’s scene references. To finalize navigation, we will create an SPI for each screen, including a string for the scene name, which will allow the screens to navigate properly among themselves.